+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/data/Babysitter.kt
@@ -1,4 +1,4 @@
-package com.example.feature_babysitter.babysitter
+package com.example.feature_babysitter.babysitter.data
 
 data class Babysitter(
     val name: String,
