+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/view/ChildListScreen.kt
@@ -1,0 +1,97 @@
+package com.example.feature_babysitter.babysitter.view
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.example.feature_babysitter.R
+import com.example.feature_babysitter.babysitter.data.BabySitterViewModel
+import com.example.feature_babysitter.babysitter.navigation.BabySitterScreen
+
+
+@Composable
+fun ChildListScreen(
+    navController: NavController,
+    viewModel: BabySitterViewModel,
+    onNavigateBack: () -> Unit
+) {
+    val children = viewModel.children
+
+    Column(modifier = Modifier.padding(top = 40.dp)) {
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp)
+                .clickable {
+                    navController.navigate(BabySitterScreen.YourChildDetails.route)
+                }
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.plus_user),
+                    contentDescription = "Child Image",
+                    modifier = Modifier
+                        .size(50.dp)
+                        .clip(CircleShape)
+                )
+                Spacer(modifier = Modifier.width(16.dp))
+
+                // Child Name
+                Text(
+                    text = "New Child",
+                    modifier = Modifier.weight(1f)
+                )
+            }
+        }
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f)
+        ) {
+            items(children) { child ->
+                ChildCard(child = child, navController, children.indexOf(child), viewModel)
+            }
+        }
+        Button(
+            onClick = {
+                navController.navigate(BabySitterScreen.FilterScreen.route)
+            },
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+                .padding(16.dp)
+        ) {
+            Text(text = "Next")
+        }
+        Button(
+            onClick = onNavigateBack,
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+                .padding(10.dp)
+        ) {
+            Text(text = "Back to All Apps")
+        }
+    }
+}
