+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/view/OrderScreen.kt
@@ -1,0 +1,175 @@
+package com.example.feature_babysitter.babysitter.view
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.LocationOn
+import androidx.compose.material3.Button
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.example.feature_babysitter.babysitter.data.BabySitterViewModel
+import com.example.feature_babysitter.babysitter.navigation.BabySitterScreen
+
+
+@Composable
+fun OrderScreen(
+    navController: NavController,
+    viewModel: BabySitterViewModel,
+    indexBabySitter: String?,
+    indexChild: String?
+) {
+    val babysitter = viewModel.babysitters[indexBabySitter!!.toInt()]
+    val child = viewModel.children[indexChild!!.toInt()]
+
+
+    Column(
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(MaterialTheme.colorScheme.primary)
+                .padding(16.dp)
+        ) {
+            // Title
+            Text(
+                text = "Order Details",
+                color = Color.White,
+                modifier = Modifier.padding(bottom = 16.dp)
+            )
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceEvenly
+            ) {
+                // Babysitter Profile
+                Column(
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Image(
+                        painter = painterResource(id = babysitter.imageResId),
+                        contentDescription = "Babysitter Image",
+                        modifier = Modifier
+                            .size(70.dp)
+                            .clip(CircleShape)
+                            .border(1.dp, Color.White, CircleShape),
+                        contentScale = ContentScale.Crop
+                    )
+                    Text(text = "Babysitter", color = Color.White)
+                    Text(text = babysitter.name, color = Color.White)
+                }
+
+                // Child Profile
+                Column(
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Image(
+                        painter = painterResource(id = child.imageResId),
+                        contentDescription = "Child Image",
+                        modifier = Modifier
+                            .size(70.dp)
+                            .clip(CircleShape)
+                            .border(1.dp, Color.White, CircleShape),
+                        contentScale = ContentScale.Crop
+                    )
+                    Text(text = "Child", color = Color.White)
+                    Text(text = child.name, color = Color.White)
+                }
+            }
+
+            // Date and Location
+            Text(
+                text = "Date",
+                color = Color.White,
+                modifier = Modifier.padding(top = 8.dp)
+            )
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Icon(
+                    imageVector = Icons.Default.LocationOn,
+                    contentDescription = "Location Icon",
+                    tint = Color.White
+                )
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(
+                    text = babysitter.location,
+                    color = Color.White
+                )
+            }
+        }
+
+        // Lower Section
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+        ) {
+            // Babysitter Cost Row
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Text(text = "Babysitter")
+                Text(text = "$${babysitter.costPerHour}/hr")
+            }
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Subtotal
+            Text(text = "Subtotal")
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Delivery Fees
+            Text(text = "Delivery Fees")
+
+            Spacer(modifier = Modifier.height(16.dp))
+            Divider(color = Color.Gray, thickness = 1.dp)
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(
+                text = "Total Amount",
+                modifier = Modifier.align(Alignment.CenterHorizontally)
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Button(
+                onClick = {
+                    navController.navigate(BabySitterScreen.BabySitterMainScreen.route)
+                },
+                modifier = Modifier.align(Alignment.CenterHorizontally)
+            ) {
+                Text(text = "Place Order")
+            }
+        }
+    }
+}
