+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/view/SearchScreen.kt
@@ -1,4 +1,4 @@
-package com.example.feature_handyman.handyman.ui
+package com.example.feature_babysitter.babysitter.view
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -14,9 +14,9 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AddCircle
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Place
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -28,38 +28,51 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.feature_handyman.R
-import com.example.feature_handyman.handyman.HandyData
+import com.example.feature_babysitter.babysitter.data.BabySitterViewModel
+import com.example.feature_babysitter.babysitter.data.Babysitter
+import com.example.feature_babysitter.babysitter.navigation.BabySitterScreen
 
-
-val handymen = listOf<HandyData>(
-    HandyData("Goku", "Earth", R.drawable.cam_placeholder, 3.0, 500, 15),
-    HandyData("Gohan", "Earth", R.drawable.cam_placeholder, 3.0, 500, 15),
-    HandyData("Goten", "Earth", R.drawable.cam_placeholder, 3.0, 500, 15),
-
-)
 
 @Composable
-fun SearchScreen(navController: NavController) {
+fun SearchScreen(
+    navController: NavController,
+    viewModel: BabySitterViewModel,
+    sort: String?,
+    max: String?,
+    min: String?
+) {
     var showPopup by remember { mutableStateOf(false) }
-    var expandedHandyData by remember { mutableStateOf<HandyData?>(null) }
+    var expandedBabysitter by remember { mutableStateOf<Babysitter?>(null) }
+    val children = viewModel.children
+    val child = children[0] // You can replace this with your dynamic child list logic
+
+    var babysitters = viewModel.babysitters
+
+    if (sort.equals("Recommend")) {
+        babysitters = viewModel.recommendSortBabysitter
+    } else if (sort.equals("Distance")) {
+        babysitters = viewModel.distanceSortBabysitter
+    } else if (sort.equals("Cost")) {
+        babysitters = viewModel.priceSortBabysitter
+    }
+
+    babysitters = babysitters.filter {
+        it.costPerHour < max!!.toDouble() && it.rating > min!!.toDouble()
+    }
 
     Column(modifier = Modifier.fillMaxSize()) {
         // Top Image
         Image(
-            painter = painterResource(id = R.drawable.cam_placeholder),
+            painter = painterResource(child.imageResId),
             contentDescription = "Top Image",
             modifier = Modifier
                 .fillMaxWidth()
                 .weight(1f)
         )
 
-        // Bar with two options: Favorites and Orders
         Row(
             modifier = Modifier
                 .fillMaxWidth()
@@ -86,23 +99,36 @@ fun SearchScreen(navController: NavController) {
             // Orders Row
             Row(
                 modifier = Modifier
-                    .clickable { showPopup = true },
+                    .clickable {
+                        navController.navigate(
+                            BabySitterScreen.BabyMapScreen.withArgs(
+                                sort.toString(),
+                                max.toString(),
+                                min.toString()
+                            )
+                        )
+                    },
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 Icon(
-                    imageVector = Icons.Default.AddCircle,
-                    contentDescription = "Orders",
+                    imageVector = Icons.Default.Place,
+                    contentDescription = "Map View ",
                     tint = Color.White
                 )
                 Spacer(modifier = Modifier.width(8.dp))
-                Text(text = "Orders", color = Color.White)
+                Text(text = "Map View", color = Color.White)
             }
         }
 
         // Show Popup Menu
-        /*if (showPopup) {
-            OrdersPopupMenu(navController, onDismiss = { showPopup = false })
-        }*/
+        if (showPopup && expandedBabysitter != null) {
+            OrdersPopupMenu(
+                navController = navController,
+                viewModel = viewModel,
+                babysitter = expandedBabysitter!!, // Passing the selected babysitter to the popup
+                onDismiss = { showPopup = false }
+            )
+        }
 
         // Bar with title "Babysitters" and filter icon
         Row(
@@ -113,29 +139,30 @@ fun SearchScreen(navController: NavController) {
             verticalAlignment = Alignment.CenterVertically,
             horizontalArrangement = Arrangement.SpaceBetween
         ) {
-            Text(text = "Babysitters")
+            Text(text = "Babysitters: ${babysitters.size} results")
             Icon(
-                modifier = Modifier.clickable { },
+                modifier = Modifier.clickable {
+                    navController.navigate(BabySitterScreen.FilterScreen.route)
+                },
                 imageVector = Icons.Default.Menu,
                 contentDescription = "Filter",
                 tint = MaterialTheme.colorScheme.primary
             )
         }
 
+        // Lazy list of babysitters
         LazyColumn(modifier = Modifier.weight(2f)) {
-            items(handymen) { handymen ->
-                HandyManCard(
-                    handyData = handymen,
-                    onClick = {  expandedHandyData  = handymen }
-                )
+            items(babysitters) { babysitter ->
+                if (!showPopup) {
+                    BabysitterCard(
+                        babysitter = babysitter,
+                        onClick = {
+                            expandedBabysitter = babysitter  // Set the selected babysitter
+                            showPopup = true                 // Show the popup
+                        }
+                    )
+                }
             }
         }
     }
-}
-
-@Preview
-@Composable
-fun SearchScreenPreview() {
-    SearchScreen(navController = NavController(LocalContext.current))
-
 }
