+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/view/YourChildDetailsScreen.kt
@@ -1,0 +1,135 @@
+package com.example.feature_babysitter.babysitter.view
+
+import android.widget.Toast
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Button
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.example.feature_babysitter.R
+import com.example.feature_babysitter.babysitter.data.BabySitterViewModel
+import com.example.feature_babysitter.babysitter.data.Child
+import com.example.feature_babysitter.babysitter.navigation.BabySitterScreen
+
+@Composable
+fun YourChildDetailsScreen(
+    navController: NavController,
+    viewModel: BabySitterViewModel,
+    onNavigateBack: () -> Unit
+) {
+    var name by remember { mutableStateOf("") }
+    var isMale by remember { mutableStateOf(true) }
+    var age by remember { mutableStateOf("") }
+    val context = LocalContext.current
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        // Title
+        Text(
+            text = "New Child Info",
+            style = MaterialTheme.typography.headlineMedium,
+            modifier = Modifier.padding(top = 35.dp)
+        )
+
+        OutlinedTextField(
+            value = name,
+            onValueChange = { input ->
+                if (input.all { it.isLetter() || it.isWhitespace() }) {
+                    name = input
+                }
+            },
+            label = { Text("Name") },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(bottom = 16.dp)
+        )
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.padding(bottom = 16.dp)
+        ) {
+            Text(text = "Gender:", modifier = Modifier.padding(end = 8.dp))
+            Checkbox(
+                checked = isMale,
+                onCheckedChange = { isMale = it },
+                modifier = Modifier.padding(end = 8.dp)
+            )
+            Text(text = if (isMale) "Male" else "Female")
+        }
+
+        // Age Entry
+        OutlinedTextField(
+            value = age,
+            onValueChange = {
+                if (it.all { char -> char.isDigit() }) {
+                    val inputAge = it.toIntOrNull() ?: 0
+                    if (inputAge in 0..17) age = it // Ensures the age is within the allowed range
+                }
+            },
+            label = { Text("Age") },
+            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(bottom = 16.dp)
+        )
+
+        Button(
+            onClick = {
+                if (name.isNullOrBlank() || age.isNullOrBlank()) {
+                    Toast.makeText(context, "Name or Age cannot be empty", Toast.LENGTH_SHORT)
+                        .show()
+                } else {
+                    val child = Child(
+                        name = name,
+                        gender = if (isMale) "Male" else "Female",
+                        age = age.toIntOrNull() ?: 0, // Convert age to Int safely
+                        imageResId = R.drawable.cam_placeholder // Use appropriate placeholder
+                    )
+                    viewModel.addChild(child) // Add child to ViewModel
+
+                    navController.navigate(BabySitterScreen.TakeAPhotoScreen.route)
+                }
+            }
+        ) {
+            Text(text = "Next")
+        }
+        Button(
+            onClick = {
+                navController.navigate(BabySitterScreen.ChildListScreen.route)
+            }
+        ) {
+            Text(text = "Back")
+        }
+        Spacer(modifier = Modifier.height(16.dp))
+
+        // Back to Home Button
+        Button(
+            onClick = onNavigateBack
+
+        ) {
+            Text(text = "Back to All Apps")
+        }
+    }
+}
